requirements.txt
async-timeout
beautifulsoup4
certifi
charset-normalizer
filelock
fsspec
g4f[all]
google-generativeai
googlesearch-python
groq
h11
httpcore
httpx
idna
Jinja2
MarkupSafe
mpmath
networkx
numpy
openai
Pillow
pydantic
pydantic_core
PySocks
python-dotenv
PyYAML
requests
soupsieve
sympy
tqdm
typing_extensions
urllib3
aiofiles
aiogram
pymupdf
aiosqlite
faster_whisper
lingua-language-detector
translate
spacy
google.genai